services/like_service/pyproject.toml
[tool.poetry]
name = "like-service"
version = "0.1.0"
description = ""
authors = ["srivatsan177"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
redis = "^5.0.1"
boto3 = "^1.34.41"
python-dotenv = "^1.0.1"
schedule = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
